fixed a bug where user could delete other users comments

+++ nc-daily-newsy/src/components/CommentDelete.jsx
@@ -3,12 +3,13 @@ import * as api from "../utils/api";
 
 class CommentDelete extends Component {
   handleDelClick = () => {
-    api.deleteCommentById(this.props.comment_id).then(() => {
-      this.props.deleteComment();
-    });
+    if (this.props.userInfo.user === this.props.author) {
+      api.deleteCommentById(this.props.comment_id).then(() => {
+        this.props.deleteComment();
+      });
+    }
   };
   render() {
-    console.log(this.props.deleteComment);
     return (
       <div>
         <button className="button" onClick={this.handleDelClick}>
